:sparkles: Add navigable field
*

Index file displays a console printed field with prompt for a direction
to move. Commands use the "a/s/d/w" directions of the days of yore.

Field class manages the state of the field. Includes methods to print
the current state of the field (`print`) and to process move commands
and potentially update the state of the board (`acceptMoveCommand`).

Coordinate class manages math related to coordinates and provides an
object data structure in which to apply consistency to coordinate use.